refactoring the code and layout
