--- handlers/schemas/bedrock-ai21.json
@@ -1,54 +1,0 @@
-{
-    "request": {
-        "defaults": {
-            "prompt": "My name is Olivier and I",
-            "numResults": 1,
-            "topKReturn": 1,
-            "maxTokens": 4096,
-            "minTokens": 0,
-            "stopSequences": [],
-            "temperature": 1.0,
-            "topP": 1.0,
-            "presencePenalty": {
-                "scale": 0,
-                "applyToWhitespaces": false,
-                "applyToNumbers": false,
-                "applyToEmojis": false,
-                "applyToPunctuations": false,
-                "applyToStopwords": false
-            },
-            "countPenalty": {
-                "scale": 0,
-                "applyToWhitespaces": false,
-                "applyToNumbers": false,
-                "applyToEmojis": false,
-                "applyToPunctuations": false,
-                "applyToStopwords": false
-            },
-            "frequencyPenalty": {
-                "scale": 0,
-                "applyToWhitespaces": false,
-                "applyToNumbers": false,
-                "applyToEmojis": false,
-                "applyToPunctuations": false,
-                "applyToStopwords": false
-            }
-        },
-        "mapping": {
-            "prompt": "$.prompt",
-            "best_of": "$.numResults",
-            "top_k": "$.topKReturn",
-            "max_new_tokens": "$.maxTokens",
-            "min_new_tokens": "$.minTokens",
-            "temperature": "$.temperature",
-            "top_p": "$.topP",
-            "stop": "$.stopSequences"            
-        }
-    },
-    "response": {
-        "mapping": {
-            "generated_text": "$.completions[0].data.text",
-            "finish_reason": "$.completions[0].finishReason"
-        }
-    }
-}
