--- handlers/schemas/bedrock-amazon-embedding.json
@@ -1,23 +1,0 @@
-{
-    "request": {
-        "defaults": {
-            "inputText": "My name is Olivier and I"
-        },
-        "mapping": {
-            "prompt": "$.inputText",
-            "max_new_tokens": "$.textGenerationConfig.maxTokenCount",
-            "temperature": "$.textGenerationConfig.temperature",
-            "top_p": "$.textGenerationConfig.topP",
-            "stop": "$.textGenerationConfig.stopSequences"
-        }
-    },
-    "response": {
-        "mapping": {
-            "generated_text": "$.results[0].outputText",
-            "finish_reason": "$.results[0].completionReason",
-            "generated_tokens_count": "$.results[0].tokenCount",
-            "input_tokens_count": "$.inputTextTokenCount",
-            "embedding": "$.embedding"
-        }
-    }
-}
