--- handlers/schemas/sagemaker-tgi.json
@@ -1,57 +1,0 @@
-{
-    "request": {
-        "defaults": {
-            "inputs": "My name is Olivier and I",
-            "parameters": {
-                "best_of": null,
-                "decoder_input_details": false,
-                "details": true,
-                "do_sample": false,
-                "max_new_tokens": 20,
-                "repetition_penalty": null,
-                "return_full_text": false,
-                "seed": null,
-                "stop": [
-                  "photographer"
-                ],
-                "temperature": null,
-                "top_k": null,
-                "top_p": null,
-                "truncate": null,
-                "typical_p": null,
-                "watermark": false
-            }
-        },
-        "mapping": {
-            "prompt": "$.inputs",
-            "stream": "$.stream",
-            "max_new_tokens": "$.parameters.max_new_tokens",
-            "repetition_penalty": "$.parameters.repetition_penalty",
-            "return_full_text": "$.parameters.return_full_text",
-            "temperature": "$.parameters.temperature",
-            "top_p": "$.parameters.top_p",
-            "top_k": "$.parameters.top_k",
-            "stop": "$.parameters.stop",
-            "best_of": "$.parameters.best_of",
-            "decoder_input_details": "$.parameters.decoder_input_details",
-            "details": "$.parameters.details",
-            "do_sample": "$.parameters.do_sample",
-            "seed": "$.parameters.seed",
-            "truncate": "$.parameters.truncate",
-            "typical_p": "$.parameters.typical_p",
-            "watermark": "$.parameters.watermark"
-        }
-    },
-    "response": {
-        "mapping": {
-            "text": "$.[0].token.text",
-            "logprobs": "$.[0].token.logprob",
-            "finish_reason": "$.[0].details.finish_reason",
-            "generated_tokens": "$.[0].details.generated_tokens",
-            "seed": "$.[0].details.seed",
-            "generated_text": "$.[0].generated_text",
-            "id": "$.[0].token.id",
-            "special": "$.[0].token.special"
-        }
-    }
-}
