+++ lambda/handlers/schemas/bedrock-anthropic.json
@@ -1,0 +1,26 @@
+{
+    "request": {
+        "defaults": {
+            "prompt": "My name is Olivier and I",
+            "max_tokens_to_sample": 4096,
+            "stop_sequences": [],
+            "temperature": 1.0,
+            "top_p": 1.0,
+            "top_k": 1
+        },
+        "mapping": {
+            "prompt": "$.prompt",
+            "max_new_tokens": "$.max_tokens_to_sample",
+            "temperature": "$.temperature",
+            "top_p": "$.top_p",
+            "top_k": "$.top_k",
+            "stop": "$.stop_sequences"
+        }
+    },
+    "response": {
+        "mapping": {
+            "generated_text": "$.completion",
+            "finish_reason": "$.stop_reason"
+        }
+    }
+}
