+++ lambda/src/requirements.txt
@@ -1,0 +1,2 @@
+boto3
+fmeval
