--- src/requirements.txt
@@ -1,2 +1,0 @@
-boto3
-fmeval
